+++ src/modules/popup/index.tsx
@@ -5,10 +5,11 @@ import {Router, Route} from 'wouter'
 import {useHashLocation} from 'wouter/use-hash-location'
 import {ROUTES} from '@/lib/routes'
 
-import {Details} from '_modules/popup/details/Details'
-import {Colors} from '_modules/popup/details/Colors'
-import {Images} from '_modules/popup/details/Images'
-import {Favorites} from '_modules/popup/details/Favorites'
+import {Details} from '@/modules/popup/Details'
+import {Colors} from '@/modules/popup/Colors'
+import {Images} from '@/modules/popup/Images'
+import {Favorites} from '@/modules/popup/Favorites'
+import {Figma} from '@/modules/popup/Figma'
 
 import '@/globals.css'
 
@@ -20,6 +21,7 @@ ReactDOM.createRoot(document.getElementById('module') as HTMLElement).render(
       <Route path={ROUTES.images} component={Images} />
 
       <Route path={ROUTES.favorites} component={Favorites} />
+      <Route path={ROUTES.figma} component={Figma} />
     </Router>
   </React.StrictMode>,
 )